Putting things in their own class files. One small fix for polygon area.
