+++ csharp/src/sdk/Geometry/Polygons.cs
